--- src/components/headline.jsx
@@ -1,34 +1,0 @@
-import React from 'react';
-
-function headline() {
-  return (
-    <header>
-      <div className="headline">
-        <h1 className="primary-text-header">I’m Steven Levoyer nice to see you!</h1>
-        <p className="primary-text-body">
-          I’m a software developer! I can help you build a product , feature or
-          website Look
-          through some of my work and experience! If you like what you see and have a project you need
-          coded, don’t
-          hestiate to contact me.
-        </p>
-        <div className="social-media">
-          <div className="supporting-text">
-            <a href="#About" className="lets-connect">LET’S CONNECT</a>
-          </div>
-          <div className="social-media-icons">
-            <ul className="header-icons">
-              <li className="sm-icons"><a href="#" /></li>
-              <li className="sm-icons"><a href="#" /></li>
-              <li className="sm-icons"><a href="#" /></li>
-              <li className="sm-icons"><a href="#" /></li>
-              <li className="sm-icons"><a href="#" /></li>
-            </ul>
-          </div>
-        </div>
-      </div>
-    </header>
-  );
-}
-
-export default headline;
