--- src/components/sidebar.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-function sidebar () {
-  return(
-    <div>
-      
-    </div>
-  )
-}
-
-export default sidebar;
